improvement: refactored structure (containers - components)
