+++ src/components/App/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { MuiThemeProvider } from '@material-ui/core';
 
-import { theme } from '../utils';
+import { theme } from '../../utils';
 import { AuthProvider } from '../AuthContext';
 import Main from '../Main';
