+++ src/containers/MapPage/MapPage.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Map from '../Map';
+import Map from '../../components/Map';
 
 class MapPage extends Component {
   render() {
